vf-global_front
- Correction des éléments suite à la réunion sur homepage.html.twig, homepage.css.
- Début de modification front pour annonces.html.twig.
- Modification annonces.html.twig suite aux instructions établies.
- Correction sommaire des éléments de style se répétant involontairement dans toutes feuilles de styles (balise 'a', 'h1', 'h2', 'p', 'input', etc... en rajoutant une classe supérieure de leur feuille de style respective)
- Suppression du doublon de la suppression des annonces périmées et création de sa fonction "deleteOldAnnonce" appelé dans la fonction homepage et annonces.
- Redirection d'une éventuelle recherche de l'utilisateur effectuer depuis le homepage.html.twig vers annonces.html.twig
