+++ public/assets/lesAnnonces.css
@@ -1,12 +1,107 @@
-/*annonces.html.twig*/
-.containerGlobaleProducts h1 {
+.globaleProducts_head {
+    display: flex;
+    flex-direction: row;
+    text-align: center;
+    width: 100%;
+    margin-top: 75px;
+}
+
+.annonces-container-title_searchProducts {
+    width: 60%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+.annonces-container-title_searchProducts h1 {
     color: #48945D;
     font-size: 2vw;
     font-family: 'Montserrat', sans-serif;
     font-weight: 700;
+    grid-column-start: 1;
+    grid-column-end: 5;
     grid-row: 1;
-    width: 90%;
+    align-self: self-end;
+    width: 80%;
     margin: 0 auto;
+}
+
+.annonces-container-search {
+    margin-top: 20px;
+    width: 55%;
+    height: 150px;
+}
+
+.annonces-search_title {
+    grid-area: titre-search;
+    margin: 5px;
+}
+
+.annonces-container-search input::-moz-placeholder {
+    color: #000;
+}
+
+.annonces-container-search input::placeholder {
+    color: #000;
+}
+
+.annonces-search_title div {
+    position: relative;
+}
+
+.autocomplete-items {
+    position: absolute;
+    border: 1px solid #d4d4d4;
+    border-bottom: none;
+    border-top: none;
+    z-index: 99;
+    /*position the autocomplete items to be the same width as the container:*/
+    top: 100%;
+    left: 0;
+    right: 0;
+}
+
+.autocomplete-items div {
+    padding: 10px;
+    cursor: pointer;
+    background-color: #fff;
+    border-bottom: 1px solid #d4d4d4;
+}
+
+/*when hovering an item:*/
+.autocomplete-items div:hover {
+    background-color: #e9e9e9;
+}
+
+/*when navigating through the items using the arrow keys:*/
+.autocomplete-active {
+    background-color: DodgerBlue !important;
+    color: #ffffff;
+}
+
+.annonces-container-btn_create {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 40%;
+}
+
+.annonces-container-btn_create a {
+    background-color: #48945D;
+    color: white;
+    padding: 10px 50px;
+    transition: background-color 0.2s;
+    box-shadow: 0 15px 10px -15px #111;
+    font-size: 18px;
+    border-radius: 25px;
+    cursor: pointer;
+    border: 2px solid #48945d;
+}
+
+.annonces-container-btn_create a:hover {
+    background-color: rgba(72, 148, 93, 0.78);
+    transition: background-color 0.2s;
 }
 
 .imageCardAnnonces {
@@ -30,7 +125,7 @@
     border-bottom-right-radius: 10px;
 }
 
-a {
+.globaleProducts_head a {
     text-decoration: none;
 }
 
@@ -94,14 +189,8 @@ a {
 
 
 .total-annonce {
-    width: 70%;
-    margin-top: 7%;
-}
-
-/*home/annonces.html.twig*/
-
-.container_imgsearchProducts {
-    display: block;
+    display: flex;
+    justify-content: center;
 }
 
 .container-btn_create_annonce {
@@ -296,7 +385,7 @@ a {
     margin: 5px 0;
 }
 
-h2 {
+.globaleProducts_head h2 {
     margin: 0;
 }
 
@@ -324,12 +413,8 @@ h2 {
 }
 
 .containerGlobaleProducts {
-    display: grid;
-    grid-area: mainProducts;
-    grid-template-columns: 1fr;
-    grid-template-areas:
-                "imgsearchProducts"
-                "contenuProducts";
+    display: flex;
+    flex-direction: column;
 }
 
 /*Formulaire de recherche*/
@@ -427,19 +512,6 @@ h2 {
                 "titre-searchProducts"
                 "bouton-searchProducts";
     align-items: center;
-}
-
-.search_title {
-    grid-area: titre-search;
-    margin: 5px;
-}
-
-.search_btn {
-    grid-area: bouton-search;
-}
-
-.titre-searchProducts {
-    grid-area: titre-searchProducts;
 }
 
 .bouton-searchProducts {
