+++ public/assets/mentions.css
@@ -18,7 +18,7 @@
   margin: 0;
 }
 
-h1 {
+.logoMentions h1 {
   font-size: max(2.8vw, 15px);
 }
