+++ public/assets/redirectionPaiement.css
@@ -27,7 +27,7 @@
     text-align: justify;
 }
 
-h2 {
+.recapLivraison h2 {
     color: #4a975f;
     font-size: 2em;
 }
