+++ templates/annonce/createAnnonce.html.twig
@@ -3,7 +3,7 @@
 {% block title %}Création d'une annonce - Plant&Flower{% endblock %}
 
 {% block js %}
-    <script defer src={{ asset('assets/Javascript/annonce.js') }}></script>
+    <script defer src={{ asset('assets/Javascript/createEditAnnonce.js') }}></script>
 {% endblock js %}
 
 {% block body %}
