+++ templates/user_profil/userProfil.html.twig
@@ -11,7 +11,7 @@
 {% block body %}
     <div class="container_profil">
         <div class="container-uP-title">
-            <h2>Bienvenue sur votre page de profil</h2>
+            <h1>Bienvenue sur votre page de profil</h1>
         </div>
         {% if app.user.owner %}
             <div class="container-uP-btn-create_annonce">
